New data model 'odoo.project.module'
This new data model is here to distinguish available upstream modules
and installed modules in a project.

It inherits from `odoo.module.branch` so it has access to all its data,
but is linked to an `odoo.project` and has its own `installed_version`
so it becomes easy to find modules that could be upgraded within a
project.