MigrationScanner: check is scanning the module is relevant
If new commits since the last scan are updating irrelevant files
regarding the collection of migration data (like PO files updated
automatically by Weblate), then we skip the scan as this is a process
that can be quite time consuming.

Even if the scan has been skipped we still push the last source and
target commits to Odoo.