MigrationScanner: no need to fetch branches
As the RepositoryScanner (executed before MigrationScanner) has already
fetched the branches of the repository, we avoid to fetch them again
so the MigrationScanner will work on the local history of commits.

This allows to speed up the processing of related jobs.
Without new commits in repositories, a scan of all Odoo + OCA repositories
now takes about ~6min instead of ~13min.