BaseScanner: apply git configuration at repository level
Instead of global level as the user running the scanner could not have
write access to ~/.gitconfig.
As such, the configuration is done after the cloning of repository,
right before the fetch and later on the checkout (which is the operation
that consumes most of the memory, the configuration is here to reduce it).

Cloning operation itself is faster and consumes less memory with
'filter=blob:none' parameter.