BaseScanner: fix '_get_last_fetched_commit' and '_checkout_branch'
We could get a local branch named `origin/17.0` colliding with the real
remote branch. In that case git doesn't complain but display a warning:

    warning: refname 'origin/17.0' is ambiguous.

To get the last fetched commit of a remote branch better to prefix the
refname with 'remotes/'.
And to checkout a remote branch, limit the refs lookup to the origin
remote.