odoo_repository: split RepositoryScanner jobs in smaller jobs
RepositoryScanner jobs could be quite long to execute on a repository
(like 'odoo/odoo'), and depending on your Odoo config or your monitoring tools,
such job could get killed before it ends up its work.

While there is no issue to get the RepositoryScanner job killed and
restarted later, it'll re-analyze the branch to scan to detect again the
modules it has to scan, and this process itself could take ages before
it re-starts the scan of the modules themselves.

This commit is splitting these two tasks:
- one job to detect modules to scan for a given repository/branch
- one job created for each module to scan

All of these jobs are depending on each other so we ensure that they are
run sequentially (one branch after another, and one module after another)
so only one job is running for a given repository at a time.
We also ensure that no more than one job can be spawned by the user from UI.

This results in a higher number of jobs, but scanning can be recovered
easily.