BaseScanner: improve clone IO performance
The use of '--filter=blob:none' was generating more disk IO operations
than without. On a network mounted volume, the performance of any git
operations decreases a lot compared to a std clone (with all blobs).

Also, to improve the time of a clone operation, we restrict it to a
single branch. If another branch is required later, it'll be fetched
automatically anyway.