[FIX] odoo_repository_migration: trigger a fetch if module not found in local source branch
Odoo database could know a module for a given repository/branch from its
last scan that the local git repo doesn't. This could happen if the
storage has been restored from an older backup for instance, right
before the migration scan of this module (pending job).

In such case, a new fetch will be done to get this module available.